+++ src/client.js
@@ -548,6 +548,13 @@ class Client extends ClientAuth {
 }
 
 /**
+ * If presence is enabled, then your presence can be set/restored.
+ *
+ * @type {Boolean} [presenceEnabled=true]
+ */
+Client.prototype.presenceEnabled = true;
+
+/**
  * Array of items to be checked to see if they can be uncached.
  *
  * @private